% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{funr}
\alias{funr}
\title{Title}
\usage{
funr(args, help_text, script_name = "funr")
}
\arguments{
\item{args}{Should always be: \code{commandArgs(trailingOnly = TRUE)}}

\item{help_text}{A simple text to be displayed describing options and usage of this interface}

\item{script_name}{Name of the script. This is used in the the help text. [funr]}
}
\description{
Title
}

