% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_home.R
\name{get_script_path}
\alias{detect_home}
\alias{get_script_path}
\title{Enables detection of the folder a script resides in with certain accuracy}
\source{
https://github.com/molgenis/molgenis-pipelines/wiki/How-to-source-another_file.R-from-within-your-R-script
}
\usage{
get_script_path()

detect_home()
}
\description{
R does not have a default way to return and use, the location of a specific script.
}

