% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/render.R
\name{render_funr}
\alias{render_funr}
\title{Render output of functions in a print friendly format}
\usage{
render_funr(x, max_rows = 100)
}
\arguments{
\item{x}{a output from \code{funr}}

\item{max_rows}{In case output of a function is a data.frame, the number
of rows to display.}
}
\description{
If the function returns with invisible, output is suppressed
}

