% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/render_out.R
\name{render_out}
\alias{render_out}
\title{Render output of functions in a print friendly format}
\usage{
render_out(x, max_rows = 100)
}
\arguments{
\item{x}{a output from \code{funr}}
}
\description{
If the function returns with invisible, output is suppressed
}

