% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_script.R
\name{sys.script}
\alias{sys.script}
\title{Get path of current script}
\source{
http://stackoverflow.com/questions/1815606/rscript-determine-path-of-the-executing-script
}
\usage{
sys.script()
}
\description{
Get path of current script
}
\examples{
\dontrun{
## cat foo.R
## library(funr);sys.script()

## source("foo.R")

## Rscript foo.R

}
}

